YTDB-523: Extract duplicate mock fixtures across 4 BTreeEngine test files
Create BTreeEngineTestFixtures with shared createMockStorage(),
injectField(), findField(), SingleValueFixture, and MultiValueFixture.
Update all 4 test files to extend the shared fixtures, eliminating
copy-pasted mock wiring code.
